Tried not resizing handles
But it did not work.

Hopefully future generations will have forgotten this, but in 2018 AE had a bug where if you ran Create ProEXR Layer Comps on a file that had a Cryptomatte and then opened the resulting Cryptomatte comp, you'd get an error saying "can't resize locked handle" and AE would be put into an unstable state, probably crashing when you quit (which isn't the worst time to crash by any means). Since I do resize handles I thought maybe I was doing something wrong, but even not resizing handles gives me the error.

So I'm inclined to say this is an AE bug. Then again, when I create arbitrary data for EXtractoR I don't have this issue.

I need the AE team to get to the bottom of the error/bug for me. Apologies in advance if it really is something I'm doing wrong, which can't be ruled out completely.